Exclude only ".git" subdirectories
Maybe excluding all top-level dotdirs is a bit too general. Now we
ignore all ".git" subdirectories which is a bit clearer and easier to
read. Also, putting the earlier in the code should make it more
performant as bigger things are excluded earlier. Not that it really
matters, but eh.