Add more canonical names
We need a canonical header for the time index, hence the inclusion of
the `"TS"` key. Fixing the `"pmp_hydro"` lookup fixes a bug and mapping
the strings at the bottom of the dictionary to the empty string is
necessary to correctly identify how useless third lines and how many
columns are used for the time index.