Compare objective values across all scenarios
Done by looping over the `scenarios` variable. Also, the "comparison.md"
include is generalized to accept the current scenario as a parameter.
Note though that this parameter has to be a String, e.g. "1", because it
pulls data out of `site.data.data[column]` while when generating the
title the key has to be an integer because it pulls data out of
`site.data.meta.scenarios`.