Change scenario metadata to be an array
It was a hash that used integer keys before, which is weird because it
basically emulates an array, if the keys are consecutive. So
transforming the scenario metadata into an array was a natural thing and
it has the following benefits:

  - The YAML file containing the metadata contains less line noise, so
    it should be more readable.
  - Iterating over the `site.data.meta.scenarios` now directly gives you
    the scenario object at each iteration step.
  - There's no ambiguity anymore on whether the keys for
    `site.data.meta.scenarios` are integers or strings.

Note that arrays are indexed zero based so now we have to subtract one
from the indexes gathered from the second level subdirectories of the
"projects" directory. Fortunately enough the `minus` filter
automatically converts its left argument to an integer, relieving us of
the duty to do so explicitly.